Remove discovery
By removing discovery, we avoid activating this plugin when we do not
mean to do so. We could create a more precise discovery rule, but
separating when we need tanka from when we would need jsonnet is not
simple.

With no specified discovery, the plugin must be specified by name. That
seems like a good option.